requirements.txt
django==3.2
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-redis==5.2.0
celery[redis]==5.2.4
django-celery-beat==2.2.1
